requirements.txt
click==8.0.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
PyJWT==2.3.0
waitress==2.0.0
Werkzeug==2.0.2
